✨ Refactor fuzz_play_game.rs and add lib.rs with fizz_buzz_fibonacci implementation
